% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_account_calc.R
\name{daily_account_calc}
\alias{daily_account_calc}
\title{Calculation of account value for backtesting asset allocation strategies}
\usage{
daily_account_calc(w, R)
}
\arguments{
\item{w}{A vector of weights}

\item{R}{An xts object with daily returns of the tickers in strat.}
}
\value{
A numeric vector with the daily value of the account.
}
\description{
\code{daily_account_calc} is a helper function used by
\code{backtest_allocation} to calculate theoretical the theoretical account
value given an initial allocation to assets. It is not intended to be called
directly by the user.
}
\details{
The function simulates the value of a theoretical account from the initial
weights and the daily returns of a set of assets.
}
