% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ETFs.R
\docType{data}
\name{ETFs}
\alias{ETFs}
\title{Daily prices and total returns for 24 ETFs.}
\format{
An object of class \code{"list"}
\describe{
 \item{Prices}{\code{xts} object with daily prices}
 \item{Returns}{\code{xts} object with daily total returns}
 \item{Description}{\code{data.frame} with information about the ETFs}
 \item{risk_free}{\code{xts} object with daily returns of U.S. Treasury bills}
}
}
\usage{
data(ETFs)
}
\description{
Data set containing daily prices and total returns for 24 exchange-traded
funds (ETFs) as well as daily returns for U.S. Treasury bills
(risk-free asset).
}
\examples{

data(ETFs)
head(ETFs$Prices)
ETFs$Description

}
\keyword{datasets}
