% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rebalance_dates.R
\name{get_rebalance_dates}
\alias{get_rebalance_dates}
\title{Portfolio rebalancing dates}
\usage{
get_rebalance_dates(dates, reb_freq, k = 1)
}
\arguments{
\item{dates}{A vector of dates}

\item{reb_freq}{Character with rebalancing frequency. Options are
\code{"days"}, \code{"weeks"}, \code{"months"}, \code{"quarters"},
and \code{"years"}}

\item{k}{An integer with number of periods to skip.}
}
\value{
A vector of dates.
}
\description{
\code{get_rebalance_dates} determines rebalancing dates based on rebalancing
frequency chosen by the user. This is a helper function used by
\code{backtest_allocation} and  is not intended to be called directly by the user.
}
