% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_variance_constraint.R
\name{min_variance_constraint}
\alias{min_variance_constraint}
\title{Full investment constraint for minimum variance portfolio optimization}
\usage{
min_variance_constraint(x, cov_mat, b, c)
}
\arguments{
\item{x}{A vector of weights.}

\item{cov_mat}{A covariance matrix, passed to optimizer}

\item{b}{A vector of desired risk budgets}

\item{c}{A scalar.}
}
\value{
A scalar with the sum of the portfolio weights.
}
\description{
\code{min_variance_constraint} is a helper function for the
\code{min_variance} function.
}
\details{
The function implements a full investment constraint necessary in the
optimization.
}
