% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/port_vol.R
\name{port_vol}
\alias{port_vol}
\title{Portfolio volatility}
\usage{
port_vol(x, cov_mat, b, c)
}
\arguments{
\item{x}{A vector of weights.}

\item{cov_mat}{A covariance matrix.}

\item{b}{A scalar parameter, passed to optimizer.}

\item{c}{A scalar parameter, passed to optimizer.}
}
\value{
A scalar with the portfolio volatility.
}
\description{
\code{port_vol} is a helper function for the \code{risk_parity} function.
}
\details{
The function calculates the volatility of a portfolio given weights and the
covariance matrix.
}
\examples{
x <- c(0.75, 0.25)
cov_mat <- matrix(c(0.25^2, 0.00125, 0.00125, 0.05^2),
                  nrow = 2, ncol = 2)
(port_vol(x, cov_mat))
}
