% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_nonlin_constraint.R
\name{rp_nonlinear_constraint}
\alias{rp_nonlinear_constraint}
\title{Nonlinear constraint for risk parity portfolio optimization}
\usage{
rp_nonlinear_constraint(x, cov_mat, b, c)
}
\arguments{
\item{x}{A vector of weights.}

\item{cov_mat}{A covariance matrix, passed to optimizer}

\item{b}{A vector of desired risk budgets}

\item{c}{A scalar.}
}
\value{
A scalar with the portfolio volatility.
}
\description{
\code{rp_nonlinear_constraint} is a helper function for the
\code{risk_parity} function.
}
\details{
The function implements a nonlinear constraint necessary in the optimization.
}
\examples{
x <- c(0.75, 0.25)
cov_mat <- matrix(c(0.25^2, 0.00125, 0.00125, 0.05^2),
                  nrow = 2, ncol = 2)
b <- c(0.5, 0.5)
c <- -10
(rp_nonlinear_constraint(x, cov_mat, b, c))
}
